Functional test upgrade.
Switched to RSH pipe for Perforce connection and upgraded to 15.1 P4D
binaries.  Made use of ClassRule for Jenkins (faster startup) and a
Rule for Perforce with separate roots (to allow for parallel test runs).
