+++ src/main/java/org/jenkinsci/plugins/p4/client/ConnectionConfig.java
@@ -1,8 +1,8 @@
 package org.jenkinsci.plugins.p4.client;
 
-import java.io.Serializable;
-
 import org.jenkinsci.plugins.p4.credentials.P4BaseCredentials;
+
+import java.io.Serializable;
 
 public class ConnectionConfig implements Serializable {
 
@@ -18,24 +18,8 @@ public class ConnectionConfig implements Serializable {
 		this.p4port = credential.getP4port();
 		this.ssl = credential.isSsl();
 		this.trust = credential.getTrust();
-		this.serverUri = toUri();
+		this.serverUri = credential.getP4JavaUri();
 		this.timeout = credential.getTimeout();
-	}
-
-	public ConnectionConfig(String p4port, boolean ssl, String trust) {
-		this.p4port = p4port;
-		this.ssl = ssl;
-		this.trust = trust;
-		this.serverUri = toUri();
-		this.timeout = 0;
-	}
-
-	private String toUri() {
-		if (ssl) {
-			return "p4javassl://" + p4port;
-		} else {
-			return "p4java://" + p4port;
-		}
 	}
 
 	public String getPort() {
