+++ src/test/java/org/jenkinsci/plugins/p4/client/PublishTest.java
@@ -1,0 +1,106 @@
+package org.jenkinsci.plugins.p4.client;
+
+import hudson.Launcher;
+import hudson.model.AbstractBuild;
+import hudson.model.BuildListener;
+import hudson.model.Cause;
+import hudson.model.FreeStyleBuild;
+import hudson.model.FreeStyleProject;
+import hudson.model.Result;
+import hudson.tasks.Builder;
+import org.jenkinsci.plugins.p4.DefaultEnvironment;
+import org.jenkinsci.plugins.p4.PerforceScm;
+import org.jenkinsci.plugins.p4.SampleServerRule;
+import org.jenkinsci.plugins.p4.populate.AutoCleanImpl;
+import org.jenkinsci.plugins.p4.populate.Populate;
+import org.jenkinsci.plugins.p4.publish.PublishNotifier;
+import org.jenkinsci.plugins.p4.publish.SubmitImpl;
+import org.jenkinsci.plugins.p4.workspace.ManualWorkspaceImpl;
+import org.jenkinsci.plugins.p4.workspace.WorkspaceSpec;
+import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+import org.jvnet.hudson.test.JenkinsRule;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.logging.Logger;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+/**
+ * Created by pallen on 01/11/2016.
+ */
+public class PublishTest extends DefaultEnvironment {
+
+	private static Logger logger = Logger.getLogger(FreeStyleTest.class.getName());
+	private static final String P4ROOT = "tmp-PublishTest-p4root";
+
+	@ClassRule
+	public static JenkinsRule jenkins = new JenkinsRule();
+
+	@Rule
+	public SampleServerRule p4d = new SampleServerRule(P4ROOT, VERSION);
+
+	@Before
+	public void buildCredentials() throws Exception {
+		createCredentials("jenkins", "jenkins", p4d);
+	}
+
+	@Test
+	public void testPublishWithPurge() throws Exception {
+		FreeStyleProject project = jenkins.createFreeStyleProject("Publish-purge");
+
+		// Create workspace
+		String client = "manual.ws";
+		String stream = null;
+		String line = "LOCAL";
+		String view = "//depot/Data/... //" + client + "/...";
+		WorkspaceSpec spec = new WorkspaceSpec(true, true, false, false, false, false, stream, line, view);
+		ManualWorkspaceImpl workspace = new ManualWorkspaceImpl("none", true, client, spec);
+
+		// Populate with P4 scm
+		Populate populate = new AutoCleanImpl();
+		PerforceScm scm = new PerforceScm(CREDENTIAL, workspace, populate);
+		project.setScm(scm);
+
+		// Create artifact files
+		project.getBuildersList().add(new CreateArtifact("artifact.1", "content"));
+		project.getBuildersList().add(new CreateArtifact("artifact.2", "content"));
+
+		// Submit artifacts
+		SubmitImpl submit = new SubmitImpl("publish", true, true, true, "3");
+		PublishNotifier publish = new PublishNotifier(CREDENTIAL, workspace, submit);
+		project.getPublishersList().add(publish);
+		project.save();
+
+		// Start build
+		Cause.UserIdCause cause = new Cause.UserIdCause();
+		FreeStyleBuild build = project.scheduleBuild2(0, cause).get();
+		assertEquals(Result.SUCCESS, build.getResult());
+
+		List<String> log = build.getLog(LOG_LIMIT);
+		assertTrue(log.contains("P4 Task: syncing files at change: 18"));
+		assertTrue(log.contains("p4 reopen -c41 -t+S3 //manual.ws/..."));
+		assertTrue(log.contains("... submitting files"));
+		assertTrue(log.contains("p4 describe -s 41"));
+	}
+
+	private static final class CreateArtifact extends Builder {
+		private final String filename;
+		private final String content;
+
+		public CreateArtifact(String filename, String content) {
+			this.filename = filename;
+			this.content = content;
+		}
+
+		@Override
+		public boolean perform(AbstractBuild<?, ?> build, Launcher launcher, BuildListener listener) throws InterruptedException, IOException {
+			build.getWorkspace().child(filename).write(content, "UTF-8");
+			return true;
+		}
+	}
+}
